Release 1.1.37
added a help text to the login screen about quick login. updated the organization request form. updated the help text in the organizations modal. added a warning to the settings screen indicating the SMS provider is out of service for now. hidden the save draft button temporarily. updated model validation for organization requests.